Split API into separate files, improve interface and add more functionality
