+++ forte/sparse_ci/sparse_operator.h
@@ -123,6 +123,13 @@ SparseOperator operator*(const SparseOperator& lhs, const SparseOperator& rhs);
 /// @return The commutator of two second quantized operators
 SparseOperator commutator(const SparseOperator& lhs, const SparseOperator& rhs);
 
+/// @brief Evaluate a factorized similarity transformation of an operator generated by a generator T
+/// @param O the operator to which the similarity transformations will be applied
+/// @param T = (T1, T2, ...) the list of operators applied in the similarity transformations
+/// @param reverse if true, apply the similarity transformations in reverse order
+/// @param screen_threshold a threshold to select which elements of the operator will be applied
+/// @details This function applies the similarity transformations in the order provided in the list
+/// T O -> ... (1 - T2) (1 - T1) O (1 + T1) (1 + T2) ...
 void sim_trans_fact_op(SparseOperator& O, const SparseOperatorList& T, bool reverse = false,
                        double screen_threshold = 1e-12);
